trashguCapstone/backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyJWT
Werkzeug
Pillow
numpy
gunicorn
requests